@@ -1,21 +1,62 @@
 __all__ = [
+    # core
     "BaseModel",
     "Field",
     "logger",
-    "cli",
+    "zyxModuleLoader",
+    # model application
+    "app",
+    # client -- core
+    "completion",
+    "embedding",
+    "image",
+    "speak",
+    "transcribe",
+    # client -- chat
+    "chat",
+    # client -- functions
     "chainofthought",
     "classify",
-    "completion",
     "code",
     "extract",
     "function",
     "generate",
-    "zyxModuleLoader",
+    # client -- data
+    "Rag",
+    # agents
+    "Agents",
+    "tools",
 ]
 
-# --- zyx ----------------------------------------------------------------
-
-from .core.ext import BaseModel, Field, zyxModuleLoader
-from .client.main import completion
-from .client.fn import classify, chainofthought, code, extract, function, generate
+# -- core --
+from .core.main import (
+    BaseModel,
+    Field,
+    zyxModuleLoader,
+)
 from loguru import logger
+
+# -- client --
+# - core
+from .client.main import completion
+from litellm.main import embedding
+from .client.multimodal import image, speak, transcribe
+
+# - chat
+from .client.chat.main import app
+from .client.chat.chat import chat
+
+# - functions
+from .client.functions.chainofthought import chainofthought
+from .client.functions.classify import classify
+from .client.functions.code import code
+from .client.functions.extract import extract
+from .client.functions.function import function
+from .client.functions.generate import generate
+
+# -- data
+from .data.rag import Rag
+
+# -- agents
+from .agents.main import Agents
+from .agents import tools as tools
